Corrected positioning of ft-sensor in urdf
The origin of the force torque sensor was set to 1.2 cm in z direction
of the ee_link, due to the assumption that the z-axis shows in the
direkten of the ft sensor/gripper. This assumption was wrong, the x-axis
of the ee_link ponts in this direction, so this was changed.
The second change is the value. In the urdf file it is described that
the adapter is 9mm height instead of 12mm. When we measure it, 9mm fits
better then 12mm.
The rotation of the force torque sensor was changed, now the connection
points into the right direction. Due to that the rotation of the gripper
had to be fixed too.