new link tool0 and prefix for ur5
We want to use pick and place, in this case we noticed that the ee_link
has another orientation as the eef of the gripper, so we inserted a new
link tool0 at the origin of the fts_toolside with the same orientation
as the s_model_tool_0. The ur5 has already a tool0, so we set a prefix
for the ur5.
The new tool0 link is the eef of the ur5 now.